Moved create_schedule into psyclone_test_utils (which then requires a __init__.py so that it can be imported from a subdirectory).
