+++ setup.py
@@ -42,7 +42,8 @@ from setuptools import setup, find_packages
 
 BASE_PATH = os.path.dirname(os.path.abspath(__file__))
 SRC_PATH = os.path.join(BASE_PATH, "src")
-PACKAGES = find_packages(where=SRC_PATH)
+PACKAGES = find_packages(where=SRC_PATH, exclude=["psyclone.tests",
+                                                  "psyclone.tests.test_files"])
 
 NAME = 'PSyclone'
 AUTHOR = ("Rupert Ford <rupert.ford@stfc.ac.uk>, "
@@ -102,7 +103,7 @@ if __name__ == '__main__':
                           'six', 'enum34 ; python_version < "3.0"'],
         extras_require={
             'doc': ["sphinx", "sphinxcontrib.bibtex", "sphinx_rtd_theme"],
-            'test': ["pytest<5.0"], # TODO: Issue 438. Fix > 5.0 broken tests.
+            'test': ["pytest<5.0"],  # TODO: Issue 438. Fix > 5.0 broken tests.
         },
         include_package_data=True,
         scripts=['bin/psyclone', 'bin/genkernelstub'],
