+++ src/psyclone/tests/dynamo0p3_basis_test.py
@@ -42,13 +42,13 @@ import os
 import pytest
 import fparser
 from fparser import api as fpapi
-from dynamo0p3_build import Dynamo0p3Build
 from psyclone.configuration import Config
 from psyclone.parse.algorithm import parse
 from psyclone.parse.utils import ParseError
 from psyclone.psyGen import PSyFactory, GenerationError, InternalError
 from psyclone.dynamo0p3 import DynKernMetadata, DynKern
-from psyclone_test_utils import print_diffs
+from .dynamo0p3_build import Dynamo0p3Build
+from .psyclone_test_utils import print_diffs
 
 # constants
 BASE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
