+++ src/psyclone/tests/dynamo0p3_builtins_test.py
@@ -43,12 +43,14 @@
 from __future__ import absolute_import, print_function
 import os
 import pytest
-from dynamo0p3_build import Dynamo0p3Build
+
 from psyclone.parse.algorithm import parse
 from psyclone.parse.utils import ParseError
 from psyclone.psyGen import PSyFactory, GenerationError
 from psyclone.configuration import Config
 from psyclone import dynamo0p3_builtins
+
+from .dynamo0p3_build import Dynamo0p3Build
 
 # constants
 BASE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
