+++ src/psyclone/tests/dynamo0p3_haloex_test.py
@@ -39,12 +39,14 @@ implementation for gh_inc dependencies using pytest. '''
 from __future__ import absolute_import
 import os
 import pytest
-from dynamo0p3_build import Dynamo0p3Build
+
 from psyclone.parse.algorithm import parse
 from psyclone.psyGen import PSyFactory
 from psyclone.dynamo0p3 import DynLoop, DynHaloExchange
 from psyclone.transformations import Dynamo0p3RedundantComputationTrans
 from psyclone.configuration import Config
+
+from .dynamo0p3_build import Dynamo0p3Build
 
 # constants
 API = "dynamo0.3"
