+++ src/psyclone/tests/profile_test.py
@@ -45,7 +45,7 @@ from psyclone.profiler import Profiler, ProfileNode
 from psyclone.psyGen import Loop, NameSpace
 from psyclone.transformations import GOceanOMPLoopTrans, OMPParallelTrans, \
     ProfileRegionTrans, TransformationError
-from psyclone_test_utils import get_invoke
+from .psyclone_test_utils import get_invoke
 
 
 # -----------------------------------------------------------------------------
